ThreadPool: Protect _workers with lock
Fixes #58.

Currently there is a race condition as ``run_thread()`` calls ``thread.start()`` before adding the given thread to ``_workers``, so if the new thread attempts to access ``_workers`` an error can occur.

Add a get_worker_thread() helper for thread-safe access. Note that for this to work, users of guikit must only access _workers via this helper (or with _workers_lock held)!