Use platformdirs to find a log directory for app
Calculating the path for an app's data correctly is a surprisingly
difficult task. The platformdirs package offers a cross-platform way of
doing this and is already an indirect dependency of the project.

Change the logging code to use platformdirs rather than open-coding a
routine to guess the app data dir.