Optionally also use app author as part of log path
platformdirs allows for passing an appauthor parameter to the functions for finding various app data folders. It appears not to be used for Linux but on Windows there *always* seems to be an (implicit) app author (if you don't specify one explicitly it assumes it's the same as the app name). Not sure about macOS!

If the user doesn't supply an app author the behaviour is as it was previously.