Don't explicitly set logging level anywhere
The default logging level -- which should be "info", provided the user
hasn't changed it -- should be fine.

Setting the logging level to "debug" unfortunately causes extremely
verbose logging for some Python modules and you can end up getting a lot
of log spam. For example, I'm trying to use the azure package in another
project and the contents of all the requests are being output to the
command line.